Remove geometric_search from geometric search files
